pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "page analiser"
authors = ["Bkorob <baka311@yandex.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.2.2"
gunicorn = "^20.1.0"
flake8 = "^5.0.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
validators = "^0.22.0"

[build-system]
requires = ["poetry-core>=1.2.2"]
build-backend = "poetry.core.masonry.api"